requirements.txt
alembic==1.0.8
appdirs==1.4.3
beautifulsoup4==4.7.1
blinker==1.4
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
coverage==4.5.3
cssselect==1.0.3
eyeD3==0.8.10
fake-useragent==0.1.11
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
fudge==1.1.1
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
libmagic==1.0
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
parse==1.12.0
Pillow==6.1.0
pipdeptree==0.13.2
pyee==6.0.0
PyMySQL==0.9.3
pyppeteer==0.0.25
pyquery==1.4.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-magic-bin==0.4.14
pytz==2018.9
requests==2.21.0
simplejson==3.16.0
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.2
tqdm==4.32.0
urllib3==1.24.3
uWSGI==2.0.18
w3lib==1.20.0
websockets==7.0
Werkzeug==0.15.2
wrapt==1.11.1
